# Enable line-only debuginfo by default.