Previously, only official builds (stable/beta/nightly) had `debuginfo-lines` enabled, but it was in conjunction with `debuginfo-only-std`, meaning that the compiler had *no* debuginfo.

In my experience, when you have a panic/crash in `rustc`, line debuginfo is *necessary* to get a readable backtrace, and we would've saved ourselves some pain and time, had we done this earlier.

At least we can try to explore and discuss the tradeoffs involved here.

r? @alexcrichton cc @rust-lang/infra @rust-lang/compiler