@RalfJung 2018-09-22
Related issue: https://github.com/rust-lang/rust/issues/52179